Refactor av soknad for ducks.
