+++ src/ducks/soknad/actions.js
@@ -1,0 +1,16 @@
+import * as Types from './types';
+
+/**
+ * Actions
+ * -----------------------------------------------------------------------
+ * Dette er action creators som returnerer Redux-klargjorte actions
+ * uten support for asynkrone kall.
+ */
+
+/* eslint-disable import/prefer-default-export */
+export function oppdaterSoknadState(dokument) {
+  return ({
+    type: Types.OPPDATER_SOKNAD,
+    dokument,
+  });
+}
