+++ src/ducks/soknad/index.js
@@ -1,0 +1,15 @@
+import * as soknadOperations from './operations';
+import * as soknadSelectors from './selectors';
+import * as soknadActions from './actions';
+import * as soknadTypes from './types';
+
+import soknadReducers from './reducers';
+
+export {
+  soknadOperations,
+  soknadSelectors,
+  soknadTypes,
+  soknadActions,
+};
+
+export default soknadReducers;
