+++ src/ducks/soknad/types.js
@@ -1,0 +1,11 @@
+/**
+ * Types
+ * ----------------------------------------------------------------------------------
+ * Dette er action types som benyttes for å indikere hvordan fagsaker skal manipuleres
+ * eller oppdateres.
+ */
+
+export const OK = 'soknad/OK';
+export const FEILET = 'soknad/FEILET';
+export const PENDING = 'soknad/PENDING';
+export const OPPDATER_SOKNAD = 'soknad/OPPDATER_SOKNAD';
